ai-social-automation-platform/backend/requirements.txt
# ==============================================
# VelocityPost.ai - Python Dependencies
# ==============================================

# Flask Framework & Extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-Session==0.5.0
Flask-Caching==2.1.0

# Database
pymongo==4.6.1
mongoengine==0.28.2
redis==5.0.1

# Background Tasks
celery==5.3.4

# Security & Authentication
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT==2.8.0
# cryptography==41.0.8
pycryptodome==3.19.0

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8
pydantic==2.5.1

# HTTP Requests & Web Scraping
requests==2.31.0
urllib3==2.1.0
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Date/Time Handling
python-dateutil==2.8.2

# AI Services
openai==1.3.9
anthropic==0.7.8

# Payment Processing
stripe==7.8.0
razorpay==1.3.0

# Data Validation & Processing
marshmallow==3.21.0
email-validator==2.1.0.post1
validators==0.22.0

# Image Processing
Pillow==10.1.0

# Text Processing & NLP
textblob==0.17.1

# Scheduling
APScheduler==3.10.4

# File Handling
python-magic==0.4.27

# Social Media APIs
tweepy==4.14.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# Production Server
gunicorn==21.2.0

# Monitoring & Logging
sentry-sdk[flask]==1.38.0
structlog==23.2.0

# System Monitoring
psutil==5.9.6

# Development Tools
flask-debugtoolbar==0.13.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code Quality
flake8==6.1.0

# Type Hints
typing-extensions==4.8.0

# Browser Automation (Optional - for platforms without APIs)
selenium==4.15.2

# Analytics & Data Processing
pandas==2.1.4
numpy==1.26.4


# API Documentation
flask-restx==1.3.0
apispec==6.3.0

# Security Headers
flask-talisman==1.1.0

# Additional Utilities
python-slugify==8.0.1
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
blinker==1.7.0

# JSON Processing
ujson==5.8.0

# UUID Generation
uuid==1.30

# Timezone Support
pytz==2023.3

# Rate Limiting
slowapi==0.1.9

# Configuration Management
dynaconf==3.2.4

# Process Management
supervisor==4.2.5

# Health Checks
py-healthcheck==1.10.1

# Memory Profiling (Development)
memory-profiler==0.61.0

# Database Migrations (if needed)
alembic==1.12.1

# Job Queue Management
rq==1.15.1

# Content Moderation
better-profanity==0.7.0

# File Upload Handling
Flask-Uploads==0.2.1

# Session Management
Flask-Session2==1.3.1

# Password Strength Validation
# password-strength==0.0.3
password_strength==0.0.3.post2


# URL Shortening
pyshorteners==1.0.1

# QR Code Generation
qrcode==7.4.2

# Excel File Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF Generation
reportlab==4.0.7

# Barcode Generation
python-barcode==0.15.1

# Color Processing
colorama==0.4.6

# Progress Bars
tqdm==4.66.1

# Configuration Validation
cerberus==1.3.5

# JWT Token Handling
python-jose==3.3.0

# OAuth2 Client
authlib==1.2.1

# WebSocket Support (Optional)
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# Caching
cachetools==5.3.2

# Async Support
asyncio-mqtt==0.11.1

# Network Utilities
netifaces==0.11.0
ping3==4.0.4

# System Information
# platform==1.0.8
distro==1.8.0

# Concurrent Processing
# concurrent-futures==3.1.1

# Regular Expression Utilities
regex==2023.10.3

# String Processing
unidecode==1.3.7

# Geolocation
geoip2==4.7.0

# Social Media Content Parsing
python-twitter==3.5
facebook-sdk==3.1.0

# Image Recognition (Optional)
face-recognition==1.3.0

# Machine Learning (Optional)
scikit-learn==1.3.2

# Time Series Processing
arrow==1.3.0

# Fuzzy String Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Encryption Utilities
passlib==1.7.4

# Network Security
pyopenssl==23.3.0

# System Resource Management
# resource==0.2.1

# Development Server
waitress==2.1.2

# Load Testing
locust==2.17.0

# Performance Monitoring
newrelic==9.2.0

# Error Handling
rollbar==0.16.3

# Environment Detection
python-dotenv[cli]==1.0.0

# Internationalization
Babel==2.13.1
Flask-Babel==4.0.0

# Phone Number Validation
phonenumbers==8.13.25

# Markdown Processing
markdown==3.5.1

# HTML Processing
bleach==6.1.0

# XML Processing
lxml==4.9.3

# CSV Processing
csvkit==1.4.0

# Command Line Interface
# argparse==1.4.0
docopt==0.6.2

# Retry Logic
tenacity==8.2.3

# Rate Limiting
limits==3.6.0

# Background Job Processing
huey==2.5.0

# Task Queue
dramatiq==1.15.0

# Message Queue
pika==1.3.2

# Event Streaming
kafka-python==2.0.2

# Search Engine
elasticsearch==8.11.0

# Graph Database
py2neo==2021.2.4


# NoSQL Database Adapters
cassandra-driver==3.29.0

# Time Series Database
influxdb-client==1.38.0

# Blockchain Integration
web3==6.12.0

# API Gateway
kong==0.1.0

# Service Discovery
python-consul==1.1.0


# Container Orchestration
kubernetes==28.1.0

# Cloud Storage
boto3==1.34.0
google-cloud-storage==2.10.0

# Message Encryption
pynacl==1.5.0

# SSL Certificate Management
certbot==2.7.4

# Backup & Recovery
# duplicity==2.1.4

# System Monitoring
statsd==4.0.1

# Performance Profiling
py-spy==0.3.14

# Memory Management
pympler==0.9

# Garbage Collection Optimization
gc-python-utils==0.0.1












# Core Flask
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
pymongo==4.6.1

# Security & Auth
PyJWT==2.8.0

# Environment
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Development
flask-debugtoolbar==0.13.1